Split non-vc extensions into their own files
